MissileCollection: Normalize public launcher interface on muzzle_location.
MissileCollection#add_from is switching on the type of its argument.
The first smell we see is that a Ship has a muzzle_location but an Alien does not.

Add a muzzle_location method to Alien, so that it shares that interface with Ship.

Note that the method #add_from is still switching on the argument type, because
the initial Missile velocities are different depending on whether the launcher
is a ship or a missile.