MissileCollection: Extract munition_velocity into launcher interface.
MissileCollection#add_from still switched on the type of its argument, to determine
the initial velocity of the missile based on its argument type. Extract this concern
into the Ship and Alien classes - this is something that they know.

Replace the conditional with a reliance on a new launcher interface method: munition_velocity.
Add munition_velocity to Ship and Alien.

But there's still a smell...